requirements.txt
mysql-connector-python==8.0.31
tksheet==5.4.1
stdiomask==0.0.6